// frontend/types.ts
export interface ClientForm {
  name: string;
  dateOfBirth: Date;
  lastName: string;
  email: string;
  phone: string;
  address: string;
  creditCardNumber: string;
  cvv: string;
  expirationDate: string;
  parish: string;
  town: string;
}

export interface ProductForm {
  name: string;
  type: string;
}